Updates and cnsolidates tess for getting node names
